Allow ff-other-file-alist to specify the extension for a new file
* lisp/find-file.el (ff-other-file-alist, ff-find-the-other-file): Add
an optional third element to the lists that make ff-other-file-alist.
If this third element is present, it is used as the extension for a
new file, when none of the other options were found.  Previously, this
was forced to be the first extension in the list, which unnecessarily
coupled search order and file-to-create.