Use buffer-local comment-continue in comment-indent-new-line
As the FIXME said, comment-indent-new-line was wiping out the value of
comment-continue even if it had been set explicitly.  Checking
local-variable-p should be sufficient to fix this.

This makes comment-indent-new-line work properly if comment-continue
can't be recalculated from comment-start, such as if comment-continue
is all whitespace.

* lisp/newcomment.el (comment-indent-new-line): Don't set
comment-continue to nil if it's local-variable-p.
* test/lisp/newcomment-tests.el
(local-comment-continue-in-comment-indent-new-line): Add a test.