Make ding terminate keyboard macros even in batch mode
ding's docstring states that it terminates keyboard macros.  But, due
to what seems to be an oversight, it does not do that while executing
in batch mode.

Generally, this means that some functions may infinite loop while run
in a keyboard macro in batch mode.

One concrete example: map-y-or-n-p will infinite-loop if run in a
keyboard macro while in batch mode.

Now ding properly terminates keyboard macros while running in batch
mode.  A test showing map-y-or-n-p behaves correctly in keyboard
macros is included.

* src/dispnew.c (bitch_at_user): Always signal while in a keyboard
macro, even if in batch mode. (bug#67836)
* test/lisp/emacs-lisp/map-ynp-tests.el (map-ynp-tests-simple-call)
(test-map-ynp-kmacro): Add.