Correctly include fixed strings before a prefix wildcard in PCM
In 63a48252306a631dc07d62d19311433c7877bd27 I fixed a bug with
the PCM implementation of substring completion, relating to the
handling of PCM wildcards.

However, this fix was incomplete.  This change completes the fix by
also including a fixed string if it appears before a `prefix'
wildcard, even if try-completion doesn't discover that fixed string
grows to a unique completion.

I discovered this bug while working on enhancements to PCM completion
related to completion-pcm-leading-wildcard.

* lisp/minibuffer.el (completion-pcm--merge-completions):
* test/lisp/minibuffer-tests.el (completion-substring-test-5):