Properly operate on current fileset revision in vc-hg-print-log
If hg log doesn't receive a -r argument, it logs starting from the
most recent commit created anywhere ("tip").  But vc-print-log is
supposed to log starting from the working revision (".").  Those are
usually the same, but not always; the current fileset might not even
exist in "tip".  Fix this by just logging "." if no START-REVISION is
passed to vc-hg-print-log.

* lisp/vc/vc-hg.el (vc-hg-print-log): If start-revision is nil,
reliably log the working revision.