Use up-to-date time in wait_reading_process_output
In ad23626030f0541e761f683661b51152128fb0b5 we changed to call
current_timespec less frequently, and instead cached it in
"now".  However, this is buggy: we only updated now when timeout
was non-zero, but timeout can be set temporarily to zero in
several places in wait_reading_process_output (such as when
checking for process status changes), which would cause us to
never update "now" and therefore never detect that a timeout
happened.

Also, this caching is wrong even in principle: since we call
Lisp code from wait_reading_process_output, substantial amounts
of time can pass, and we'd be left using an outdated "now" and
incorrectly not time out.  Also, the caching isn't a useful
performance optimization: current_timespec is fast.  So just get
rid of it.

* src/process.c (wait_reading_process_output): Stop caching
current_timespec.